Total damages, double everywhere, fix event queue

+++ include/Mana.h
@@ -17,9 +17,9 @@
 #define NB_TOWERS_MAX 35
 
 typedef struct {
-    uint64_t quantity;
-    uint64_t level;
-    uint64_t max;
+    double quantity;
+    double level;
+    double max;
 } Mana;
 
 /**
@@ -36,7 +36,7 @@ Mana init_mana(void);
  * @param mana Address of the Mana structure
  * @param quantity quantity of mana to add
  */
-void add_mana(Mana *mana, uint64_t quantity);
+void add_mana(Mana *mana, double quantity);
 
 /**
  * @brief Increases the mana level, if possible
@@ -94,7 +94,7 @@ bool mana_fuse_gem(Mana *mana, Error *error);
  * @param error Address of the Error to modify it there's not enough mana
  * @return bool true if buying a gem was possible
  */
-bool mana_buy_gem(Mana *mana, uint64_t level, Error *error);
+bool mana_buy_gem(Mana *mana, double level, Error *error);
 
 /**
  * @brief Gives the amount of mana required in order to buy a new tower
@@ -102,7 +102,7 @@ bool mana_buy_gem(Mana *mana, uint64_t level, Error *error);
  * @param nb_tower the number of towers currently in the game
  * @return int the amount of mana required to buy a new tower
  */
-uint64_t mana_required_tower(int nb_tower);
+double mana_required_tower(int nb_tower);
 
 /**
  * @brief Adds mana if the time_left before the newt wave is superior to zero
